+++ examples/example-star.c
@@ -45,9 +45,6 @@ int main() {
   // Calculated quantities ------------------------------------------------->
   double az, el;                    // calculated azimuth and elevation at observing site
 
-  // Intermediate variables we'll use -------------------------------------->
-  struct timespec unix_time;        // Standard precision UNIX time structure
-
 
   // We'll print debugging messages and error traces...
   novas_debug(NOVAS_DEBUG_ON);
@@ -107,11 +104,8 @@ int main() {
   // -------------------------------------------------------------------------
   // Set the astrometric time of observation...
 
-  // Get the current system time, with up to nanosecond resolution...
-  clock_gettime(CLOCK_REALTIME, &unix_time);
-
-  // Set the time of observation to the precise UTC-based UNIX time
-  if(novas_set_unix_time(unix_time.tv_sec, unix_time.tv_nsec, LEAP_SECONDS, DUT1, &obs_time) != 0) {
+  // Set the time of observation to the current UTC-based UNIX time
+  if(novas_set_current_time(LEAP_SECONDS, DUT1, &obs_time) != 0) {
     fprintf(stderr, "ERROR! failed to set time of observation.\n");
     return 1;
   }